+++ src/simmate/configuration/django/settings.py
@@ -218,6 +218,15 @@ INSTALLED_APPS = [
     # (i.e. AJAX calls can be made to update the views)
     "django_unicorn",
     #
+    # Django simple history lets you track history of changes (and who made
+    # those changes) for a given model. This is important for models that users
+    # interact with and edit in the UI
+    "simple_history",
+    #
+    # Django contrib comments let you track comments on a model and handles
+    # moderation / user / date features for you.
+    "django_comments",
+    #
     # Other third-party apps/tools to consider. Note that some of these don't
     # need to be installed apps while some also request different setups.
     #   django-ratelimit
@@ -327,6 +336,8 @@ MIDDLEWARE = [
     "django.contrib.auth.middleware.AuthenticationMiddleware",
     "django.contrib.messages.middleware.MessageMiddleware",
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
+    # This is for tracking WHO changes a model (attaches User)
+    "simple_history.middleware.HistoryRequestMiddleware",
 ]
 
 # "core" here is based on the name of my main django folder
@@ -429,9 +440,18 @@ CRISPY_TEMPLATE_PACK = "bootstrap4"
 
 # These settings help configure djangorestframework and our REST API
 REST_FRAMEWORK = {
-    # The REST framework has both authentication AND permission classes. By
-    # default I allow anything, but may revisit these if the user-base ever grows:
-    "DEFAULT_AUTHENTICATION_CLASSES": [],
+    # The REST framework has both authentication AND permission classes. Adding
+    # the auth classes ensure we have user data available when rendering the
+    # template. Then I allow anyone (anonymous and signed in users) to access
+    # the REST APIs. Note, even though this is allow-all, we still have the
+    # "RequireLoginMiddleware" applied elsewhere if the admin choses.
+    "DEFAULT_AUTHENTICATION_CLASSES": [
+        "rest_framework.authentication.BasicAuthentication",
+        "rest_framework.authentication.SessionAuthentication",
+    ],
+    "DEFAULT_PERMISSION_CLASSES": [
+        "rest_framework.permissions.AllowAny",
+    ],
     # Because we have a massive number of results for different endpoints,
     # we want to set results to be paginated by 25 results per page. This
     # way we don't have to set a page limit for every individual endpoint. Note