+++ src/simmate/website/core/urls.py
@@ -110,6 +110,9 @@ urlpatterns = [
     # Django-unicorn urls
     path("unicorn/", include("django_unicorn.urls")),
     #
+    # Django-contrib-comments urls
+    path("comments/", include("django_comments.urls")),
+    #
     # Custom Simmate apps (if present)
     path(route="apps/", view=views.apps, name="apps"),
     *get_app_urls(),