+++ docs/change_log.md
@@ -79,6 +79,7 @@ There is one key exception to the rules above -- and that is with `MAJOR`=0 rele
 - add Microsoft allauth support to sign in use external account
 - allow "login required" access to server via evironment variables (disabled by default)
 - allow "internal only" access to server via evironment variables (disabled by default)
+- add `django-simple-history` support to track user changes on specific models
 
 **Fixes**